requirements.txt
llama-index
google-generativeai
llama-index-llms-gemini
Pypdf
python-dotenv
IPython
llama-index-embeddings-gemini
streamlit>=1.30.0
numpy
pandas
scikit-learn
PyPDF2>=3.0.0
llama-index-core>=0.10.0